utils/hdx_api.py
import shutil
import os
import logging

from hdx.utilities.easy_logging import setup_logging
from hdx.hdx_configuration import Configuration
from hdx.data.dataset import Dataset


HDX_SITE = 'prod'
USER_AGENT = 'MapAction'


setup_logging()
logger = logging.getLogger()
Configuration.create(hdx_site=HDX_SITE, user_agent=USER_AGENT, hdx_read_only=True)


def query_api(hdx_address, directory, dataset_names=None):
    if dataset_names is None:
        dataset_names = []
    dataset = Dataset.read_from_hdx(hdx_address)
    resources = dataset.get_resources()
    filenames = {}
    for resource in resources:
        if resource['name'] in dataset_names or not dataset_names:
            _, path = resource.download()
            filename = os.path.basename(path)
            shutil.move(path, os.path.join(directory, filename))
            filenames[resource['name']] = filename
            logging.info(f'Saved \"{resource["name"]}\" to \"{directory}/{filename}\"')
    return filenames
