requirements.txt
numpy==1.18.2
pandas==1.0.3
geopandas==0.7.0
snakemake==5.13.0
hdx-python-api==4.3.5
requests==2.23.0
PyYAML==5.3.1